[Core/GUI] Complete STL-to-Qt container migration (std::list, std::map, std::set, std::string)
Migrate all remaining STL containers to Qt-native equivalents across the
entire codebase (core library + GUI), excluding 3rd-party interfaces
(Vector DBC).

- std::list → QList (protocol drivers, CAN monitor buffers, system views)
- std::map → QMap (exports, protocol drivers, project GUI, user settings)
- std::set → QSet with qHash() for custom types (CAN signals, node IDs)
- std::string → QString (HALC config, datapool content, user settings)
- Add default constructor to C_OscComMessageLoggerOsySysDefConfig for QMap
- Fix QMap::erase(key) → QMap::remove(key) in C_OscComAutoSupport
- Fix QMap::emplace() → QMap::insert() in C_SyvComMessageMonitor

Two std::set instances intentionally kept (private inner class without
qHash, persistent iterator as class member).

+++ opensyde_tool/src/can_monitor/message_generator/C_CamGenTableModel.cpp
@@ -548,7 +548,7 @@ QVariant C_CamGenTableModel::data(const QModelIndex &orc_Index,
         switch (e_Col) // lint !e788 //not all columns explicitly handled
         {
         case eID:
-          c_Retval = C_Uti::h_GetValueAsHex(0ULL);
+          c_Retval = C_Uti::h_GetValueAsHex(static_cast<uint64_t>(0ULL));
           break;
         case eDLC:
           c_Retval = 0;
@@ -577,9 +577,9 @@ QVariant C_CamGenTableModel::data(const QModelIndex &orc_Index,
         {
         case eID:
           if (pc_Message->GetExtended()) {
-            c_Retval = C_Uti::h_GetValueAsHex(0x1FFFFFFFULL);
+            c_Retval = C_Uti::h_GetValueAsHex(static_cast<uint64_t>(0x1FFFFFFFULL));
           } else {
-            c_Retval = C_Uti::h_GetValueAsHex(0x7FFULL);
+            c_Retval = C_Uti::h_GetValueAsHex(static_cast<uint64_t>(0x7FFULL));
           }
           break;
         case eDLC:
